Add space detection to prefer larger spaces
This still needs a lot of work, but it's a start at using a
breadth-first search to detect which move results in the largest space
to move into. This is missing movement of other snakes, so it's
limited in it's usefulness, but should prevent moving into a single
space just to get food and dying immediately.