Add logging to a file for each game
This is part of an effort to get to the bottom of test cases that fail
in the actual game but when the postion is downloaded they work fine.

After logging the actual move data, it seems like your snake isn't
always first in the list. I think this assumption is being made
somewhere in the grid logic.