Prefer H2H over wall
Somtimes if there are H2H moves that would lose, the snake sees
that there are no good moves if the opponent plays well, and then
just runs into the wall (or whatever the 'default') move is.

Add a 'lucky' move so that the snake will try for one of the moves
that could kill it, but might not if the other snake goes a
different way.