Adjust eval of health, food, and length
Since food was scored higher the hungrier the snake is, it would prefer not
eating and having "potential" food, and so would starve in stupid ways or go
in circles around the food. Now it gets a flat multiple for all three that are
balanced to make the tests pass.