Yay! does not crash (new sibling eval)
Try to evaluate a set of moves from the point of view of all the
snakes, eliminating the worst move for a snake one at a time
until there is only one move left from a set of possibilities.

Might still be buggy, also has lots of print statements left in
for this version.